Remove Sinoptico links and add JSON export/import on home
