+++ index.html
@@ -9,7 +9,6 @@
 <body>
   <nav class="main-nav">
     <a href="#/home">Inicio</a>
-    <a href="#/sinoptico">Sinóptico</a>
     <a href="sinoptico-editor.html">Editar Sinóptico</a>
     <a href="#/amfe">AMFE</a>
     <a href="#/settings">Ajustes</a>
