+++ js/views/sinoptico.js
@@ -1,4 +1,4 @@
-import { getAll, remove, exportJSON, importJSON } from '../dataService.js';
+import { getAll, remove } from '../dataService.js';
 import { createSinopticoEditor } from '../editors/sinopticoEditor.js';
 
 export async function render(container) {
@@ -7,9 +7,6 @@ export async function render(container) {
       <button id="sin-edit">Editar</button>
       <button id="btnNuevoCliente">Nuevo cliente</button>
       <button id="sin-delete">Borrar</button>
-      <button id="sin-export">Exportar</button>
-      <input id="sin-import-file" type="file" accept="application/json" hidden>
-      <button id="sin-import">Importar</button>
     </div>
     <table id="sinoptico">
       <thead>
@@ -53,24 +50,6 @@ export async function render(container) {
     for (const item of items) await remove('sinoptico', item.id);
   });
 
-  container.querySelector('#sin-export').addEventListener('click', async () => {
-    const json = await exportJSON();
-    const a = document.createElement('a');
-    a.href = URL.createObjectURL(new Blob([json], { type: 'application/json' }));
-    a.download = 'sinoptico.json';
-    a.click();
-    URL.revokeObjectURL(a.href);
-  });
-
-  const fileInput = container.querySelector('#sin-import-file');
-  container.querySelector('#sin-import').addEventListener('click', () => fileInput.click());
-
-  fileInput.addEventListener('change', async ev => {
-    const file = ev.target.files[0];
-    if (!file) return;
-    const text = await file.text();
-    await importJSON(text);
-  });
 }
 
 // expose editor factory for renderer.js
